feat(testing): split old tests into smaller files
