+++ xieffect/test/education/interaction_test.py
@@ -1,14 +1,13 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Iterator
-from random import randint, shuffle  # noqa: DUO102
 
 from flask.testing import FlaskClient
 from pytest import mark
 from werkzeug.test import TestResponse
 
 from __lib__.flask_fullstack import check_code, dict_equal
-from .knowledge_test import MODULES_PER_REQUEST
+from .module_list_test import MODULES_PER_REQUEST
 
 
 @mark.order(500)
@@ -118,41 +117,3 @@ def test_module_opener(client: FlaskClient):  # relies on module#5 and module#9 
     assert check_code(client.get("/modules/9/open/")) == {"id": 8}
     check_code(client.get("/modules/9/points/9/"))
     assert check_code(client.get("/modules/9/open/")) == {"id": 9}
-
-
-@mark.order(540)
-def test_reply_and_results(client: FlaskClient):  # relies on module#7
-    module = check_code(client.get("/modules/7/"))
-    assert module["type"] == "test"
-    assert "map" in module
-
-    length: int = len(module["map"])
-
-    def create_reply():
-        return {
-            "right-answers": (right := randint(0, 10)),
-            "total-answers": (total := (randint(1, 5) if right == 0 else randint(right, right * 2))),
-            "answers": {str(k): int(k) for k in range(randint(right, total))}
-        }
-
-    replies: list[dict] = [create_reply() for _ in range(length)]
-
-    point_ids: list[int] = list(range(length))
-    shuffle(point_ids)
-
-    for point in point_ids:
-        data = check_code(client.get(f"/modules/7/points/{point}/reply/"))
-        assert isinstance(data, dict) and len(data) == 0
-        check_code(client.get(f"/modules/7/points/{point}/"))
-        data = check_code(client.get(f"/modules/7/points/{point}/reply/"))
-        assert isinstance(data, dict) and len(data) == 0
-
-        reply = replies[point]
-        assert check_code(client.post(f"/modules/7/points/{point}/reply/", json=reply)).get("a", False)
-        assert check_code(client.get(f"/modules/7/points/{point}/reply/")) == reply["answers"]
-
-    point_id: int | None = None
-    for i, reply in enumerate(check_code(client.get("/modules/7/results/"))["result"]):
-        assert dict_equal(reply, replies[i], "right-answers", "total-answers", "answers")
-        assert point_id is None or reply["point-id"] > point_id
-        point_id = reply["point-id"]
