+++ xieffect/test/education/result_test.py
@@ -1,12 +1,50 @@
 from __future__ import annotations
 
 from collections.abc import Callable, Iterator
-from random import randint  # noqa: DUO102
+from random import randint, shuffle  # noqa: DUO102
 
 from flask.testing import FlaskClient
 from pytest import mark
 
 from __lib__.flask_fullstack import check_code, dict_equal
+
+
+@mark.order(540)
+def test_reply_and_results(client: FlaskClient):  # relies on module#7
+    module = check_code(client.get("/modules/7/"))
+    assert module["type"] == "test"
+    assert "map" in module
+
+    length: int = len(module["map"])
+
+    def create_reply():
+        return {
+            "right-answers": (right := randint(0, 10)),
+            "total-answers": (total := (randint(1, 5) if right == 0 else randint(right, right * 2))),
+            "answers": {str(k): int(k) for k in range(randint(right, total))}
+        }
+
+    replies: list[dict] = [create_reply() for _ in range(length)]
+
+    point_ids: list[int] = list(range(length))
+    shuffle(point_ids)
+
+    for point in point_ids:
+        data = check_code(client.get(f"/modules/7/points/{point}/reply/"))
+        assert isinstance(data, dict) and len(data) == 0
+        check_code(client.get(f"/modules/7/points/{point}/"))
+        data = check_code(client.get(f"/modules/7/points/{point}/reply/"))
+        assert isinstance(data, dict) and len(data) == 0
+
+        reply = replies[point]
+        assert check_code(client.post(f"/modules/7/points/{point}/reply/", json=reply)).get("a", False)
+        assert check_code(client.get(f"/modules/7/points/{point}/reply/")) == reply["answers"]
+
+    point_id: int | None = None
+    for i, reply in enumerate(check_code(client.get("/modules/7/results/"))["result"]):
+        assert dict_equal(reply, replies[i], "right-answers", "total-answers", "answers")
+        assert point_id is None or reply["point-id"] > point_id
+        point_id = reply["point-id"]
 
 
 @mark.order(545)
